+++ fc/src/time.cc
@@ -2,8 +2,8 @@
 //from https://github.com/idealvin/coost/blob/master/src/time.cc
 //from https://github.com/matt-42/lithium/blob/master/libraries/http_server/http_server/timer.hh
 namespace fc {
-  void sleep::h(long long h) { std::this_thread::sleep_for(std::chrono::hours(h)); };
-  void sleep::m(long long m) { std::this_thread::sleep_for(std::chrono::minutes(m)); };
+  void sleep::h(int h) { std::this_thread::sleep_for(std::chrono::hours(h)); };
+  void sleep::m(int m) { std::this_thread::sleep_for(std::chrono::minutes(m)); };
   void sleep::s(long long s) { std::this_thread::sleep_for(std::chrono::seconds(s)); };
   void sleep::ms(long long ms) { std::this_thread::sleep_for(std::chrono::milliseconds(ms)); };
   void sleep::us(long long us) { std::this_thread::sleep_for(std::chrono::microseconds(us)); };
